requirements.txt
google-api-python-client==2.143.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-storage==2.18.2
loguru==0.7.2
openai==1.43.0
openpyxl==3.1.5
pandas==2.2.2
replicate==0.32.1
streamlit==1.38.0
streamlit-ext==0.1.10
extra-streamlit-components==0.1.71
httpx==0.27.0
